Refactor: step 2: remove server side api proxy
